feat: receive screens implemented
