+++ src/features/Wallet/index.ts
@@ -1,20 +1,19 @@
 import ProfileScreen from './Profile'
 import WalletScreen from './Wallet'
-import WalletSendScreen from './WalletSend'
-import WalletSendConfirmScreen from './WalletSendConfirm'
-import WalletReceiveScreen from './WalletReceive'
-import WalletReceiveConfirmScreen from './WalletReceiveConfirm'
+import SendScreen from './Send'
+import SendConfirmScreen from './SendConfirm'
+import ReceiveScreen from './Receive'
+import RequestScreen from './Request'
 import ScanScreen from './Scan'
 import TransactionHistoryScreen from '../TransactionHistory/TransactionHistory'
-import WalletSendConfirmScreenTest from '../TestScreen'
+
 export {
   ProfileScreen,
   WalletScreen,
-  WalletSendScreen,
-  WalletSendConfirmScreen,
-  WalletSendConfirmScreenTest,
-  WalletReceiveScreen,
-  WalletReceiveConfirmScreen,
+  SendScreen,
+  SendConfirmScreen,
+  ReceiveScreen,
+  RequestScreen,
   ScanScreen,
   TransactionHistoryScreen,
 }
